Convert PDBS to Multi-Element Character Strings

Convert a bio3d multi-structure pdbs object to a multi-element character vector that can be used as input for NGLVieweR and r3dmol.

Usage

pdbs2string(pdbs, collapse = TRUE)

Arguments

pdbs

a multi-structure pdbs object as obtained from pdbaln(), read.fasta.pdb(), etc.

collapse

logical, if TRUE a single element vector is returned. If FALSE a multi-element vector with a element per pdb structure is returned. The later is required for setting distinct viewing options per structure - such as user defined colors (e.g. one per structure) etc.

Value

a character vector of structure data.

See also

Author

Barry Grant, bjgrant@ucsd.edu

Examples

  pth <- "~/Desktop/courses/BIMM143/class10/pdbs/split_chain/"
  files <- list.files(path=pth, full.names = TRUE)
  pdbs <- bio3d::pdbaln(files, fit=TRUE, exefile="msa")
#> Reading PDB files:
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1AKE_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1E4V_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1E4Y_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//3HPR_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//3X2S_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4K46_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4NP6_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4X8H_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4X8M_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//5EJE_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6HAM_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6HAP_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6RZE_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6S36_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8BQF_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8PVW_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8Q2B_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8RJ9_A.pdb
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> ...   PDB has ALT records, taking A only, rm.alt=TRUE
#> ..   PDB has ALT records, taking A only, rm.alt=TRUE
#> ....   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> ..   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .
#> 
#> Extracting sequences
#> 
#> pdb/seq: 1   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1AKE_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 2   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1E4V_A.pdb 
#> pdb/seq: 3   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1E4Y_A.pdb 
#> pdb/seq: 4   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//3HPR_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 5   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//3X2S_A.pdb 
#> pdb/seq: 6   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4K46_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 7   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4NP6_A.pdb 
#> pdb/seq: 8   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4X8H_A.pdb 
#> pdb/seq: 9   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4X8M_A.pdb 
#> pdb/seq: 10   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//5EJE_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 11   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6HAM_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 12   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6HAP_A.pdb 
#> pdb/seq: 13   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6RZE_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 14   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6S36_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 15   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8BQF_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 16   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8PVW_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 17   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8Q2B_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 18   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8RJ9_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE

 NGLVieweR::NGLVieweR( bio3d::pdbs2pdb(pdbs), format="pdb") |>
    NGLVieweR::addRepresentation("cartoon")

 # Or more simpley...
 view.pdbs(pdbs)

 # Trace, tube, line, cartoon, ball+stick
 view.pdbs(pdbs, representation = "trace")

 view.pdbs(pdbs, cols = c("red","blue") )
#> Warning: Not enough distinct cols for each structure, recycling

 view.pdbs(pdbs, colorScheme = "residueindex")