Quick molecular viewing of bio3d PCA results

Quick molecular viewing of bio3d PCA results

Usage

view.pca(
  pc,
  colorScheme = "residueindex",
  representation = "cartoon",
  backgroundColor = "white",
  ...
)

view.nma(
  pc,
  colorScheme = "residueindex",
  representation = "cartoon",
  backgroundColor = "white",
  ...
)

Arguments

pc

an object of class "pca" as obtained with function pca.xyz or pca.

colorScheme

keyword based coloring used only if cols input is NULL. Possible values include "residueindex", "modelindex", "sstruc", "bfactor", "chainid", "chainindex", "atomindex", and "occupancy".

representation

the representation style, useful values are "line", "tube", "cartoon", "trace", "backbone", and "ball+stick".

backgroundColor

a single element color vector that set the display area background color.

...

additional arguments passed to and from functions (e.g. to function write.pdb).

Value

an NGLVieweR display object that can be displayed or further added to using NGLVieweR::addRepresentation() and friends.

Functions

view.nma(): Quick viewing of NMA modes

Author

Barry Grant, bjgrant@ucsd.edu

Examples

#pdb <- bio3d::read.pdb("6s36")
#n <- bio3d::nma(pdb)
#view.nma(n)
#view.nma(n, colorScheme = "model")