Interactive PDBS Viewing

Quick interactive multi-structure ensemble bio3d pdbs object viewing using the NGLVieweR package.

Usage

view.pdbs(
  pdbs,
  cols = NULL,
  colorScheme = NULL,
  representation = "cartoon",
  backgroundColor = "white"
)

Arguments

pdbs

a multi-structure pdbs object as obtained from pdbaln(), read.fasta.pdb(), etc.

cols

a vector of colors, typically one entry per structure. If NULL then the output of vmd.colors() is used.

colorScheme

if not NULL then this over-rides the previous cols input argument. Possible values include "residueindex", "modelindex", "sstruc", "bfactor", "chainid", "chainindex", "atomindex", "occupancy"

representation

the representation style, useful values are line, tube, cartoon, trace, and backbone, ball+stick.

backgroundColor

set the display area background color.

Value

an NGLVieweR display object that can be displayed or further added to using NGLVieweR::addRepresentation() and friends.

Details

Just like the related view.pdb() function the objective here is to speed up the inspection of multi structure ensembles without having to write out many aligned and superposed individual PDB files for viewing in a external molecular viewer (like PyMol, Mol-star, or VMD - all of which are excellent but often overkill for a quick inspection) or write many lines of NGLVieweR code that loop through all structures in a given pdbs object.

See also

Author

Barry Grant, bjgrant@ucsd.edu

Examples

  pth <- "~/Desktop/courses/BIMM143/class10/pdbs/split_chain/"
  files <- list.files(path=pth, full.names = TRUE)
  pdbs <- bio3d::pdbaln(files, fit=TRUE, exefile="msa")
#> Reading PDB files:
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1AKE_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1E4V_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1E4Y_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//3HPR_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//3X2S_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4K46_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4NP6_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4X8H_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4X8M_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//5EJE_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6HAM_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6HAP_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6RZE_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6S36_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8BQF_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8PVW_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8Q2B_A.pdb
#> /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8RJ9_A.pdb
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> ...   PDB has ALT records, taking A only, rm.alt=TRUE
#> ..   PDB has ALT records, taking A only, rm.alt=TRUE
#> ....   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> ..   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .   PDB has ALT records, taking A only, rm.alt=TRUE
#> .
#> 
#> Extracting sequences
#> 
#> pdb/seq: 1   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1AKE_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 2   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1E4V_A.pdb 
#> pdb/seq: 3   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//1E4Y_A.pdb 
#> pdb/seq: 4   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//3HPR_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 5   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//3X2S_A.pdb 
#> pdb/seq: 6   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4K46_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 7   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4NP6_A.pdb 
#> pdb/seq: 8   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4X8H_A.pdb 
#> pdb/seq: 9   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//4X8M_A.pdb 
#> pdb/seq: 10   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//5EJE_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 11   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6HAM_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 12   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6HAP_A.pdb 
#> pdb/seq: 13   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6RZE_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 14   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//6S36_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 15   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8BQF_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 16   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8PVW_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 17   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8Q2B_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE
#> pdb/seq: 18   name: /Users/barry/Desktop/courses/BIMM143/class10/pdbs/split_chain//8RJ9_A.pdb 
#>    PDB has ALT records, taking A only, rm.alt=TRUE

 view.pdbs(pdbs, representation = "cartoon")

 # Trace, tube, line, cartoon, ball+stick
 view.pdbs(pdbs, representation = "trace")

 view.pdbs(pdbs, cols = c("red","blue") )
#> Warning: Not enough distinct cols for each structure, recycling

 # Perhaps this should be the default?
 view.pdbs(pdbs, colorScheme = "residueindex")